feat(reporting): serialize reports differently depending on norm requirements (#2284)
* chore(reporting): organize reports by norm in bucket

As the number of reports increases, it will be good to keep them as organized as possible,
this is a simple first stab at it.  Also provides a simple way to
use a different serialization format depending on the norm.

* feat(reporting): serialize El Salvador reports differently depending on norm

Some norms require xml, some require pipe delimited text files.
This PR implements these requirements for the normas we're generating.

* fix: include nrp_51 reports